+++ src/mocks/handlers.ts
@@ -2,11 +2,11 @@ import { API_BASE_URL } from '@/constants/api-constants'
 import { http, HttpResponse } from 'msw'
 import { bookmarkedRecruitmentsHandlers } from '@/mocks/handlers/bookmarked-recruitments-handler'
 import { appliedRecruitmentsHandlers } from '@/mocks/handlers/applied-recruitments-handler'
-import { userInformationHandlers } from '@/mocks/handlers/user-information-handler'
 import { chatHandlers } from '@/mocks/handlers/chat-handler'
 import { bookmarkedLecturesHandlers } from '@/mocks/handlers/bookmarked-lectures-handler'
 import { notificationHandlers } from '@/mocks/handlers/notification-handler'
 import { completedStudiesHandlers } from '@/mocks/handlers/completed-study-handler'
+import { authHandlers } from '@/mocks/handlers/auth-handler'
 
 const getTestMSW = http.get(`${API_BASE_URL}/`, () => {
   return HttpResponse.text('msw is working!')
@@ -16,7 +16,7 @@ export const handlers = [
   getTestMSW,
   ...bookmarkedRecruitmentsHandlers,
   ...appliedRecruitmentsHandlers,
-  ...userInformationHandlers,
+  ...authHandlers,
   ...chatHandlers,
   ...bookmarkedLecturesHandlers,
   ...notificationHandlers,
